returning function to workspace service
